+++ config/config.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 )
 
-const Version = "1.3.1"
+const Version = "1.3.2"
 
 type Config struct {
 	Token               string